Cleaned up the IR implementation, fixed some bugs.

+++ source/intermediate_representation/codegen/memory/memory.h
@@ -31,7 +31,7 @@ namespace ir {
 	};
 
 	struct mem {
-		i32 index;
+		u8 index;
 		scale scale;
 		i32 displacement;
 	};
